Add files via upload
project to design and implement a new information retrieval system that will replace an existing information retrieval system (Baseline system) of a University library.
The implemented information retrieval system enables the user to index, search and retrieve abstracts from academic publications based on queries.
The application was implemented in C# using Lucene.NET 3.0.3. and Visual studio 2015 was used as the IDE to implement the system.  Text processing techniques were used to implement the application.
Wordnet was used a lexical database to perform the task query expansion.
This project was done as a part of my MSc in IT